requirements.txt
absl-py==1.4.0
alabaster==0.7.13
annexremote==1.6.0
antspynet==0.2.3
antspyx==0.3.8
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astor==0.8.1
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bids-validator==1.12.0
black==23.7.0
bleach==6.0.0
blinker==1.4
boto==2.49.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
chart-studio==1.1.0
ci-info==0.3.0
citeproc-py==0.6.0
click==8.1.5
cloudpickle==2.2.1
codecov==2.1.13
comm==0.1.3
contourpy==1.1.0
coverage==4.5.4
cryptography==3.4.8
cycler==0.11.0
Cython==3.0.0
datalad==0.19.2
dcm2niix
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dipy==1.7.0
distro==1.7.0
dm-tree==0.1.8
docopt==0.6.2
docutils==0.20.1
duecredit==0.9.2
einops==0.6.1
etelemetry==0.3.0
exceptiongroup==1.1.2
executing==1.2.0
fasteners==0.18
fastjsonschema==2.17.1
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.41.0
formulaic==0.5.2
fqdn==1.5.1
fuzzywuzzy==0.18.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.56.2
h5py==3.9.0
httplib2==0.20.2
humanize==4.7.0
idna==3.4
imageio==2.31.1
imagesize==1.4.1
importlib-metadata==4.6.4
iniconfig==2.0.0
intensity-normalization==2.2.4
interface-meta==1.3.0
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
iso8601==2.0.0
isodate==0.6.1
isoduration==20.11.0
jaraco.classes==3.3.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.23.0
keras==2.13.1
keyring==23.5.0
keyrings.alt==4.2.0
kiwisolver==1.4.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
libclang==16.0.6
lightning==2.0.7
looseversion==1.3.0
lxml==4.9.3
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.1
monai==1.2.0
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.5
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.7.2
nbformat==5.9.1
nbsphinx==0.9.2
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==3.1
nibabel==3.2.2
nilearn==0.10.1
nipy==0.5.0
nipype==1.8.6
nitime==0.10.1
notebook==7.0.0
notebook_shim==0.2.3
num2words==0.5.12
numpy==1.24.3
oauthlib==3.2.0
opt-einsum==3.3.0
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
paramiko==3.2.0
parso==0.8.3
pathspec==0.11.1
patool==1.12
patsy==0.5.3
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.9.1
plotly==5.15.0
pluggy==1.2.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.23.4
prov==2.0.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybids==0.16.1
pycparser==2.21
pydantic==1.10.11
pydicom==2.4.2
pydot==1.4.2
Pygments==2.15.1
PyJWT==2.3.0
pymedio==0.2.14
PyNaCl==1.5.0
pyparsing==2.4.7
pytest==7.4.0
pytest-cov==2.10.1
pytest-doctestplus==0.13.0
pytest-env==0.8.2
pytest-timeout==2.1.0
python-dateutil==2.8.2
python-gitlab==3.15.0
python-json-logger==2.0.7
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
rdflib==6.3.2
referencing==0.30.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
rsa==4.9
scikit-fuzzy==0.4.2
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
SecretStorage==3.3.1
Send2Trash==1.8.2
SimpleITK==2.2.1
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.4.1
Sphinx==7.0.1
sphinx-argparse==0.4.0
sphinxcontrib-apidoc==0.3.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==2.0.19
stack-data==0.6.2
statsmodels==0.14.0
sympy==1.12
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-probability==0.20.1
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.2.0
tifffile==2023.7.18
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
traits==6.3.2
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
wadllib==1.3.6
wandb==0.15.9
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
widgetsnbextension==4.0.8
wrapt==1.15.0
xvfbwrapper==0.2.9
zipp==1.0.0
